Movida conversion de nube de puntos a grilla a su propio archivo y agregada getPossiblePercentages
