+++ Competencias/Robocup_2021/Equipo/FinalCode/Analysis.py
@@ -167,85 +167,7 @@ class PathFinder:
         return []
     
 
-class ConverterPoint:
 
-    def __init__(self, startingTime, maxTime, definitiveThreshold):
-        self.definitive = False
-        self.count = 0
-        self.maxTime = maxTime
-        self.definitiveThreshold = definitiveThreshold
-        self.startingTime = startingTime
-        self.time = 0
-    
-    def update(self, time):
-        self.time = time - self.startingTime
-        if not self.definitive:
-            if self.count >= self.definitiveThreshold:
-                self.definitive = True
-            elif self.time > self.maxTime:
-                self.count = 0
-                self.startingTime = self.time
-
-    def increaseCount(self):
-        self.count += 1
-
-class PointCloudToTileConverter:
-
-    __tilesDict = {"wallRight": np.array([[0, 0, 0, 0, 1, 1],
-                                          [0, 0, 0, 0, 1, 1],
-                                          [0, 0, 0, 0, 1, 1],
-                                          [0, 0, 0, 0, 1, 1],
-                                          [0, 0, 0, 0, 1, 1],
-                                          [0, 0, 0, 0, 1, 1]]),
-
-                    "wallLeft": np.array([[1, 1, 0, 0, 0, 0],
-                                          [1, 1, 0, 0, 0, 0],
-                                          [1, 1, 0, 0, 0, 0],
-                                          [1, 1, 0, 0, 0, 0],
-                                          [1, 1, 0, 0, 0, 0],
-                                          [1, 1, 0, 0, 0, 0]]),
-                    
-                    "wallUp": np.array([[1, 1, 1, 1, 1, 1],
-                                        [1, 1, 1, 1, 1, 1],
-                                        [0, 0, 0, 0, 0, 0],
-                                        [0, 0, 0, 0, 0, 0],
-                                        [0, 0, 0, 0, 0, 0],
-                                        [0, 0, 0, 0, 0, 0]]),
-                    
-                    "wallDown": np.array([[0, 0, 0, 0, 0, 0],
-                                          [0, 0, 0, 0, 0, 0],
-                                          [0, 0, 0, 0, 0, 0],
-                                          [0, 0, 0, 0, 0, 0],
-                                          [1, 1, 1, 1, 1, 1],
-                                          [1, 1, 1, 1, 1, 1]])
-                  }
-
-    def __init__(self, size, initialTime, maxPointTime, pointPermanenceThreshold):
-
-        self.maxPointTime = maxPointTime
-        self.pointPermanenceThresh = pointPermanenceThreshold
-        self.possibleGrid = self.makeGrid(initialTime)
-        self.size = size
-
-    def makeGrid(self, time):
-        row = []
-        for _ in self.size:
-            row.append(ConverterPoint(time, self.maxPointTime, self.pointPermanenceThresh))
-        grid = []
-        for _ in self.size:
-            grid.append(row.copy())
-        return np.array(grid)
-
-
-    def update(self, time, pointCloud):
-        for point in pointCloud:
-            self.possibleGrid[point[0], point[1]].increaseCount()
-            self.possibleGrid[point[0], point[1]].update(time)
-
-
-class PointCloudToNodeGridConverter:
-    def __init__(self):
-        self.pointCloud = []
 
 
 chunk1 = np.array([[0, 1],
